+++ InscryptionAPI/Card/TribeManager.cs
@@ -142,7 +142,7 @@ public static class TribeManager
     }
     private static Texture2D MakePlaceholderCardback(Texture2D tribeIcon)
     {
-        Texture2D emptyCardback = TextureHelper.GetImageAsTexture("empty_rewardCardBack.png", typeof(TribeManager).Assembly);
+        Texture2D emptyCardback = TextureHelper.GetImageAsTexture("empty_rewardCardBack.png", InscryptionAPIPlugin.APIAssembly);
         if (tribeIcon == null) // if no tribe icon, return the empty card texture
             return emptyCardback;