+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.23.2
+- Fixed error when trying to create cards that evolve/transform in 4 or more turns
+- Fixed Fledgling and Transformer sigil appearing as black boxes when a card evolves/transform in 4 or more turns
+- Evolve and Transformer icons now show support cards that evolve in 4 or more turns
+- Added extension methods for PlayableCard and CardInfo: ProvidesBlueGem, ProvidesGreenGem, ProvidesOrangeGem
+
 # 2.23.1
 - Fixed non-CardModificationInfo shields not breaking
 - Fixed ShieldGeneratorItem not correclty updating visuals